+++ glide-core/src/lib.rs
@@ -15,6 +15,6 @@ pub use client::ConnectionRequest;
 pub mod cluster_scan_container;
 pub mod request_type;
 pub use telemetrylib::{
-    GlideOpenTelemetry, GlideOpenTelemetryConfigBuilder, GlideOpenTelemetrySignalsExporter,
-    GlideSpan, Telemetry, DEFAULT_FLUSH_SIGNAL_INTERVAL_MS, DEFAULT_TRACE_SAMPLE_PERCENTAGE,
+    DEFAULT_FLUSH_SIGNAL_INTERVAL_MS, DEFAULT_TRACE_SAMPLE_PERCENTAGE, GlideOpenTelemetry,
+    GlideOpenTelemetryConfigBuilder, GlideOpenTelemetrySignalsExporter, GlideSpan, Telemetry,
 };